BEGIN FileArchiveManager
This begins a series of commits with the goal of separating FileArchive
management from Database files' management; and to store archive names
in the database to be able to have more than FileArchive and also unify
Trash, Sandbox, etc as FileArchives in the same place.
These are mostly un-compilable. Also the previous commit also contained
some comments about these but it was not 'officially' started, yet.